+++ services/src/main/java/org/keycloak/services/clientregistration/AbstractClientRegistrationProvider.java
@@ -26,6 +26,8 @@ import org.keycloak.models.ClientModel;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.ModelDuplicateException;
 import org.keycloak.models.RealmModel;
+import org.keycloak.models.UserManager;
+import org.keycloak.models.UserModel;
 import org.keycloak.models.utils.ModelToRepresentation;
 import org.keycloak.models.utils.RepresentationToModel;
 import org.keycloak.protocol.saml.SamlConfigAttributes;
@@ -46,6 +48,9 @@ import org.keycloak.timer.TimerProvider;
 import org.keycloak.validation.ValidationUtil;
 
 import javax.ws.rs.core.Response;
+
+import static java.lang.Boolean.FALSE;
+import static java.lang.Boolean.TRUE;
 
 /**
  * @author <a href="mailto:sthorger@redhat.com">Stian Thorgersen</a>
@@ -159,6 +164,13 @@ public abstract class AbstractClientRegistrationProvider implements ClientRegist
 
         if (!client.getClientId().equals(rep.getClientId())) {
             throw new ErrorResponseException(ErrorCodes.INVALID_CLIENT_METADATA, "Client Identifier modified", Response.Status.BAD_REQUEST);
+        }
+
+        UserModel serviceAccount = this.session.users().getServiceAccount(client);
+        if (TRUE.equals(rep.isServiceAccountsEnabled()) && serviceAccount == null) {
+            new ClientManager(new RealmManager(session)).enableServiceAccount(client);
+        } else if (serviceAccount != null && FALSE.equals(rep.isServiceAccountsEnabled())) {
+                new UserManager(session).removeUser(session.getContext().getRealm(), serviceAccount);
         }
 
         RepresentationToModel.updateClient(rep, client);